+++ src/Changelog.jl
@@ -13,7 +13,7 @@ using AbstractTrees
 using OrderedCollections: OrderedDict
 import CommonMark as CM
 
-VERSION >= v"1.11.0-DEV.469" && eval(Meta.parse("public parsefile, VersionInfo, SimpleLog, generate"))
+VERSION >= v"1.11.0-DEV.469" && eval(Meta.parse("public parsefile, VersionInfo, SimpleChangelog, generate"))
 
 # generate Documenter changelogs and links
 include("generate.jl")
@@ -25,8 +25,8 @@ using .CommonMarkMarkdownASTInterop: md_convert
 # Convert MarkdownAST tree to our own tree
 include("heading_tree.jl")
 
-# SimpleLog and VersionInfo types, as well as API entrypoints
-include("SimpleLog.jl")
+# SimpleChangelog and VersionInfo types, as well as API entrypoints
+include("SimpleChangelog.jl")
 
 # Tree traversal and parsing code
 include("parse_changelog.jl")