+++ test/runtests.jl
@@ -7,4 +7,4 @@ test_path(filename) = joinpath(pkgdir(Changelog), "test", "test_changelogs", fil
 
 include("generate.jl")
 include("parse_changelog.jl")
-include("SimpleLog.jl")
+include("SimpleChangelog.jl")