+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- ## [Unreleased]
 
 ### Added
-- Public types `SimpleLog` and `VersionInfo` which capture a simple in-memory representation of a changelog
-- `Base.parse(SimpleLog, input)` and public function `parsefile` for parsing `SimpleLog`s from files and other representations. -->
+- Public types `SimpleChangelog` and `VersionInfo` which capture a simple in-memory representation of a changelog
+- `Base.parse(SimpleChangelog, input)` and public function `parsefile` for parsing `SimpleChangelog`s from files and other representations. -->
 
 ## [v1.1.0] - 2023-11-13
 ### Added